// dashboard/src/api/models/auditEvent.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Scheduler Public API (v1)
 * OpenAPI spec version: 1.3.0
 */
import type { AuditEventMetadata } from './auditEventMetadata';

export interface AuditEvent {
  eventId: string;
  /** @nullable */
  actorId?: string | null;
  action: string;
  targetType: string;
  /** @nullable */
  targetId?: string | null;
  /** @nullable */
  metadata?: AuditEventMetadata;
  createdAt: string;
}